requirements.txt
audio-recorder-streamlit==0.0.8
beautifulsoup4==4.11.2
deeplake==3.6.5
elevenlabs==0.2.18
langchain==0.0.208
openai==0.27.8
python-dotenv==1.0.1
streamlit==1.23.1
streamlit-chat==0.0.2.2
tiktoken==0.4.0
